refactor(cmd/index): move index subcommands from cmd to internal/index
